+++ packages/lib/src/parent/setParent.ts
@@ -1,8 +1,11 @@
 import { action } from "mobx"
 import { enqueuePendingAction } from "../action/pendingActions"
+import { getModelMetadata } from "../model/getModelMetadata"
+import { isModel } from "../model/utils"
 import { _BaseModel } from "../model/_BaseModel"
 import { attachToRootStore, detachFromRootStore } from "../rootStore/attachDetach"
 import { isRootStore } from "../rootStore/rootStore"
+import { clone } from "../snapshot/clone"
 import { isTweakedObject } from "../tweaker/core"
 import { failure, inDevMode, isPrimitive } from "../utils"
 import {
@@ -21,34 +24,44 @@ import { fastGetParentPath, fastGetRoot, ParentPath } from "./path"
  */
 export const setParent = action(
   "setParent",
-  (
-    value: any,
-    parentPath: ParentPath<any> | undefined,
-    indexChangeAllowed: boolean,
+  ({
+    value,
+    parentPath,
+    indexChangeAllowed,
+    isDataObject,
+    cloneIfApplicable,
+  }: {
+    value: any
+    parentPath: ParentPath<any> | undefined
+    indexChangeAllowed: boolean
     isDataObject: boolean
-  ): void => {
+    cloneIfApplicable: boolean
+  }): any => {
     if (isPrimitive(value)) {
-      return
+      return value
     }
 
     if (inDevMode()) {
+      if (indexChangeAllowed && cloneIfApplicable) {
+        throw failure(
+          "assertion failed: 'indexChangeAllowed' and 'cloneIfApplicable' cannot be set at the same time"
+        )
+      }
       if (typeof value === "function" || typeof value === "symbol") {
         throw failure(`assertion failed: value cannot be a function or a symbol`)
       }
       if (!isTweakedObject(value, true)) {
         throw failure(`assertion failed: value is not ready to take a parent`)
       }
-      if (parentPath) {
-        if (!isTweakedObject(parentPath.parent, true)) {
-          throw failure(`assertion failed: parent is not ready to take children`)
-        }
+      if (parentPath && !isTweakedObject(parentPath.parent, true)) {
+        throw failure(`assertion failed: parent is not ready to take children`)
       }
     }
 
     if (isDataObject) {
       dataObjectParent.set(value, parentPath!.parent)
       // data object will proxy to use the actual parent model for child/parent stuff
-      return
+      return value
     }
 
     initializeObjectChildren(value)
@@ -64,13 +77,25 @@ export const setParent = action(
       }
     }
 
-    const oldParentPath = fastGetParentPath(value)
+    let oldParentPath = fastGetParentPath(value)
     if (parentPathEquals(oldParentPath, parentPath)) {
-      return
+      return value
     }
 
     if (isRootStore(value)) {
       throw failure("root stores cannot be attached to any parents")
+    }
+
+    // value type models should be cloned when they are about to be assigned to a new parent
+    // and they had none previously
+    if (
+      cloneIfApplicable &&
+      parentPath?.parent &&
+      isModel(value) &&
+      getModelMetadata(value).valueType
+    ) {
+      value = clone(value, { generateNewIds: true })
+      oldParentPath = fastGetParentPath(value)
     }
 
     if (oldParentPath && parentPath) {
@@ -78,19 +103,19 @@ export const setParent = action(
         // just changing the index
         objectParents.set(value, parentPath)
         reportParentPathChanged(value)
-        return
+        return value
       } else {
         throw failure("an object cannot be assigned a new parent when it already has one")
       }
     }
 
-    const removeFromOldParent = () => {
+    const attachToNewParent = () => {
+      // detach from old
       if (oldParentPath?.parent) {
         removeObjectChild(oldParentPath.parent, value)
       }
-    }
 
-    const attachToNewParent = () => {
+      // attach to new
       objectParents.set(value, parentPath)
       if (parentPath?.parent) {
         addObjectChild(parentPath.parent, value)
@@ -101,9 +126,9 @@ export const setParent = action(
     if (value instanceof _BaseModel) {
       const oldRoot = fastGetRoot(value)
       const oldRootStore = isRootStore(oldRoot) ? oldRoot : undefined
-      removeFromOldParent()
 
       attachToNewParent()
+
       const newRoot = fastGetRoot(value)
       const newRootStore = isRootStore(newRoot) ? newRoot : undefined
 
@@ -119,8 +144,9 @@ export const setParent = action(
         })
       }
     } else {
-      removeFromOldParent()
       attachToNewParent()
     }
+
+    return value
   }
 )