+++ packages/lib/src/snapshot/SnapshotterAndReconcilerPriority.ts
@@ -1,0 +1,10 @@
+/**
+ * @ignore
+ * @internal
+ */
+export enum SnapshotterAndReconcilerPriority {
+  Array,
+  Frozen,
+  Model,
+  PlainObject,
+}