+++ packages/lib/src/snapshot/reconcileFrozenSnapshot.ts
@@ -1,6 +1,7 @@
 import { Frozen, frozen, isFrozenSnapshot } from "../frozen/Frozen"
 import { registerReconciler } from "./reconcileSnapshot"
 import type { SnapshotInOfFrozen } from "./SnapshotOf"
+import { SnapshotterAndReconcilerPriority } from "./SnapshotterAndReconcilerPriority"
 
 function reconcileFrozenSnapshot(value: any, sn: SnapshotInOfFrozen<Frozen<any>>): Frozen<any> {
   // reconciliation is only possible if the target is a Frozen instance with the same data (by ref)
@@ -11,7 +12,7 @@ function reconcileFrozenSnapshot(value: any, sn: SnapshotInOfFrozen<Frozen<any>>
   return frozen(sn.data)
 }
 
-registerReconciler(2, (value, sn) => {
+registerReconciler(SnapshotterAndReconcilerPriority.Frozen, (value, sn) => {
   if (isFrozenSnapshot(sn)) {
     return reconcileFrozenSnapshot(value, sn)
   }