+++ packages/lib/src/tweaker/TweakerPriority.ts
@@ -1,0 +1,10 @@
+/**
+ * @ignore
+ * @internal
+ */
+export enum TweakerPriority {
+  Model,
+  Array,
+  PlainObject,
+  Frozen,
+}