+++ packages/lib/src/tweaker/tweakFrozen.ts
@@ -4,6 +4,7 @@ import { setParent } from "../parent/setParent"
 import { setInternalSnapshot } from "../snapshot/internal"
 import { tweakedObjects } from "./core"
 import { registerTweaker } from "./tweak"
+import { TweakerPriority } from "./TweakerPriority"
 
 /**
  * @ignore
@@ -13,7 +14,14 @@ export function tweakFrozen<T extends Frozen<any>>(
   parentPath: ParentPath<any> | undefined
 ): T {
   tweakedObjects.set(frozenObj, undefined)
-  setParent(frozenObj, parentPath, false, false)
+  setParent({
+    value: frozenObj,
+    parentPath,
+    indexChangeAllowed: false,
+    isDataObject: false,
+    // a frozen is not a value-type
+    cloneIfApplicable: false,
+  })
 
   // we DON'T want data proxified, but the snapshot is the data itself
   setInternalSnapshot(frozenObj, { [frozenKey]: true, data: frozenObj.data })
@@ -21,7 +29,7 @@ export function tweakFrozen<T extends Frozen<any>>(
   return frozenObj as any
 }
 
-registerTweaker(5, (value, parentPath) => {
+registerTweaker(TweakerPriority.Frozen, (value, parentPath) => {
   if ((value as any) instanceof Frozen) {
     return tweakFrozen(value as Frozen<any>, parentPath)
   }